refactor: extract community profile hook
